chore(developer): rename write-compiled-keyboard.ts to kmw-compiler.ts
